requirements.txt
streamlit
pandas
langchain
langchain-openai
python-dotenv
